requirements.txt
asgiref==3.2.10
bootstrap4==0.1.0
cffi==1.14.0
Django==3.0.8
django-braces==1.14.0
django-ckeditor==5.9.0
django-filter==2.3.0
django-js-asset==1.2.2
django-taggit==1.3.0
gunicorn==20.0.4
Pillow==7.2.0
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
